created MVC classes in different files
