+++ billing.html
@@ -7,7 +7,10 @@
     <link rel="stylesheet" href="billing.css">
     <link rel="stylesheet" href="navbar.css">
     <script src="https://unpkg.com/htmlincludejs"></script>
-    <script type="text/javascript" src="/mvc_billing.js" defer></script>
+    <!-- <script type="text/javascript" src="/mvc_billing.js" defer></script> -->
+    <script type="module" src="./billing_model.mjs"></script>
+    <script type="module" src="./billing_view.mjs"></script>
+    <script type="module" src="./billing_controller.mjs"></script>
     <title>Your Shopping Cart</title>
 </head>
 <body>
