+++ homepage.html
@@ -7,7 +7,9 @@
     <link rel="stylesheet" href="/homepage.css">
     <link rel="stylesheet" href="/navbar.css">
     <script src="https://unpkg.com/htmlincludejs"></script>
-    <script type="text/javascript" src="/mvc_homescript.js" defer></script>
+    <script type="module" src="./home_model.mjs" defer></script>
+    <script type="module" src="./home_view.mjs" defer></script>
+    <script type="module" src="./home_controller.mjs" defer></script>
     <title>Everything delivered in minutes | blinkit</title>
 </head>
 <body>
